+++ Assets/Prefabs/InventoryRepair.prefab
@@ -45,9 +45,10 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 77fe3f69804125e48bfecb357b757beb, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  id: 
+  overrideNameId: 
   InventoryShape:
   - {x: 0, y: 0}
+  ShouldBeVisibleInInventory: 1
   texture: {fileID: 2800000, guid: e2995cd5f50f3ce4ca64a695b0d33230, type: 3}
 --- !u!1 &9099563949683748931
 GameObject:
@@ -156,6 +157,6 @@ MonoBehaviour:
   m_GameObject: {fileID: 9099563949683748931}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: b65305321d2d1f649a182036a8e16e7d, type: 3}
+  m_Script: {fileID: 11500000, guid: 715841e1a64987b4eb1f9b276195a203, type: 3}
   m_Name: 
   m_EditorClassIdentifier:
