+++ Assets/deCrawl/Scrips/Primitives/WorldCamera.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 715841e1a64987b4eb1f9b276195a203
+guid: a8cd2547df57fb6479cbfdd02aee57c1
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
