+++ Assets/deCrawl/Scrips/World/AbstractLootableManifestation.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 715841e1a64987b4eb1f9b276195a203
+guid: b65305321d2d1f649a182036a8e16e7d
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
